extract and type utils and partitions object
